test: split UI suites and fake domains
