+++ tests/ui/app-shell.test.tsx
@@ -1,8 +1,14 @@
-import "./helpers/module-mocks"
+import "./helpers/module-mocks/feature-detection"
+import "./helpers/module-mocks/pwa"
+import "./helpers/module-mocks/preferences"
+import "./helpers/module-mocks/qr-scanner"
+import "./helpers/module-mocks/key-records"
+import "./helpers/module-mocks/pq-records"
 import { render, screen, waitFor, within } from "@testing-library/react"
 import userEvent from "@testing-library/user-event"
 import { afterEach, beforeEach, describe, expect, it } from "vitest"
-import { fakeFeatures, warmQrReader } from "./helpers/fakes"
+import { fakeFeatures } from "./helpers/fakes/feature-detection"
+import { warmQrReader } from "./helpers/fakes/qr-scanner"
 import { setTestOnlineStatus } from "./helpers/network"
 import { renderApp, resetUi } from "./helpers/render-app"
 import { translate } from "@/i18n/messages"
