+++ tests/ui/helpers/fakes/boot.ts
@@ -1,0 +1,3 @@
+import { vi } from "vitest"
+
+export const armMaintenanceToken = vi.fn(async () => undefined)
