+++ tests/ui/helpers/module-mocks/pwa.ts
@@ -1,0 +1,9 @@
+import { vi } from "vitest"
+import * as fakes from "../fakes/pwa"
+
+vi.mock("virtual:pwa-register/react", () => ({
+  useRegisterSW: fakes.useFakeRegisterSW,
+}))
+vi.mock("@/hooks/use-register-sw", () => ({
+  useDefaultRegisterSW: fakes.useFakeRegisterSW,
+}))
