+++ tests/ui/offline-ack.test.tsx
@@ -1,4 +1,7 @@
-import "./helpers/module-mocks"
+import "./helpers/module-mocks/feature-detection"
+import "./helpers/module-mocks/pwa"
+import "./helpers/module-mocks/preferences"
+import "./helpers/module-mocks/qr-scanner"
 import { useEffect } from "react"
 import { act, render, screen, waitFor } from "@testing-library/react"
 import userEvent from "@testing-library/user-event"
@@ -17,7 +20,9 @@ import { OfflineAckShell } from "@/components/offline-ack-shell"
 import { LanguageProvider } from "@/i18n"
 import { translate } from "@/i18n/messages"
 import { decision, response } from "../helpers/boot-fixtures"
-import { fakeFeatures, getPreferences, useFakeRegisterSW } from "./helpers/fakes"
+import { fakeFeatures } from "./helpers/fakes/feature-detection"
+import { useFakeRegisterSW } from "./helpers/fakes/pwa"
+import { getPreferences } from "./helpers/fakes/preferences"
 import { setTestOnlineStatus, stubReachabilityFetch } from "./helpers/network"
 import {
   expectLanguageField,
